#226: comic
