+++ src/domain/comic/create.ts
@@ -2,8 +2,9 @@
 import createImage from '../image/create';
 import eraseImage from '../image/erase';
 import type ComicData from './data/ComicData';
-import createData from './data/create';
+import createData from './data/createData';
 import { IMAGE_TYPE } from './definitions/constants.js';
+import insert from './repository/insert';
 
 export default async function create(imageDataUrl: string, structure?: string): Promise<ComicData>
 {
@@ -12,7 +13,9 @@ export default async function create(imageDataUrl: string, structure?: string): 
     try
     {
         image = await createImage(IMAGE_TYPE, imageDataUrl);
-        data = await createData(image.id, structure);
+        data = createData(image.id, structure);
+
+        await insert(data);
 
         return data;
     }
