+++ src/domain/comic/erase.ts
@@ -1,7 +1,7 @@
 
 import eraseImage from '../image/erase';
-import eraseData from './data/erase';
-import retrieveData from './data/retrieve';
+import eraseData from './repository/erase';
+import retrieveData from './repository/retrieve';
 
 export default async function erase(id: string): Promise<void>
 {
