+++ src/domain/comic/repository/erase.ts
