+++ src/domain/comic/repository/mapFrom.ts
@@ -1,0 +1,13 @@
+
+import { type RecordData } from '^/integrations/database/module';
+
+import ComicData from '../data/ComicData';
+
+export default function mapFrom(record: RecordData): ComicData
+{
+    return new ComicData(
+        record.id as string,
+        record.imageId as string,
+        record.structure as string
+    );
+}
