+++ src/domain/comic/repository/retrieve.ts
@@ -1,13 +1,13 @@
 
 import database from '^/integrations/database/module';
 
+import type ComicData from '../data/ComicData';
 import { RECORD_TYPE } from '../definitions/constants';
-import type ComicData from './ComicData';
-import mapRecord from './mapRecord';
+import mapFrom from './mapFrom';
 
 export default async function retrieve(id: string): Promise<ComicData>
 {
     const record = await database.readRecord(RECORD_TYPE, id);
 
-    return mapRecord(record);
+    return mapFrom(record);
 }
